src/backend_pgsql.c: Fix compiler warnings and crash due to UB.
Newer releases of GCC complain about pointless/inappropriate usage of
functions like strncpy and strncat.

Additionally, since <crypt.h> was not included, the crypt() function was
not properly defined and would consequently result in the application
crashing. This is likely a result of changes introduced in GLibc as I
did not encounter this issue until I had upgraded to 2.29.